chore: move function generator and validation to utils package
